+++ notebooks/intro.clj
@@ -1,48 +1,49 @@
 ;; # Clay
 
-;; ![quaternary clay in Estonia](https://upload.wikimedia.org/wikipedia/commons/2/2c/Clay-ss-2005.jpg)
-;;
+^:kindly/hide-code?
+(with-meta
+  [:img
+   {:style {:width "300px"}
+    :src "https://upload.wikimedia.org/wikipedia/commons/2/2c/Clay-ss-2005.jpg"
+    :alt "quaternary clay in Estonia"}]
+  {:kindly/kind :kind/hiccup})
 ;; (credit: [Wikimedia Commons](https://commons.wikimedia.org/wiki/File:Clay-ss-2005.jpg))
 
 ;; ## What is it?
 
-;; [Clay](https://github.com/scicloj/clay) is a Clojure tool for data visualization and literate programming, offering a dynamic workflow compabible with popular visual tools & libraries such as [Portal](https://github.com/djblue/portal), [Clerk](https://github.com/nextjournal/clerk), and [Scittle](https://github.com/babashka/scittle). Here, by visual tools we mean tools for data visualization and literate programming.
+;; [Clay](https://github.com/scicloj/clay) is a Clojure tool for data visualization and literate programming,
+;; which is compatible with the [Kindly](https://github.com/scicloj/kindly) convention.
 
 ;; It is one of the fruits of our explorations at the [visual-tools-group](https://scicloj.github.io/docs/community/groups/visual-tools/).
-
-;; This document has been created using Clay.
 
 ;; ### Goals
 
 ;; * *Easily* **explore & share** things for others to *easily* **pick & use**.
 
-;; Small examples, library tutorials, research reports, etc., should be easily shareable as code that renders to visual documents. Not only sharing things should be obvious, but also using the shared things.
+;;   * Small examples, library tutorials, research reports, etc., should be shareable as code that renders to visual documents, that others can use.
 
 ;; * In **common** use cases, have **compatible** experiences at the relevant tools.
 
-;; Common visualizations such as images, tables, Vega/Vega-Lite plots, plain Clojure data structures, etc., should be describeable in a way that can just work in all relevant tools (e.g., Portal, Clerk) without any code change between tools.
-
-;; * In **all** use cases, be able to use **the best** tools.
-
-;; Achieving such compatibility across tools in common cases should not limit the use of the tools in any case. In all cases, it should be possible to use the best of the amazing features of tools such as Portal & Clerk.
+;;   * Common visualizations such as images, tables, Vega/Vega-Lite plots, plain Clojure data structures, etc., should be describeable in a way that other visual tools can interpret correctly.
 
 ;; * **Flow** with the REPL.
 
-;; We wish to encourage user interactions that flow naturally with the typical use of Clojure in editors and REPLs, editing and evaluating Clojure forms.
+;;   * We wish to encourage user interactions that flow naturally with the typical use of Clojure in editors and REPLs, editing and evaluating Clojure forms.
 
 ;; ### The stack
 
 ;; Clay is part of a stack of libraries seeking easy experience with common data-centric tasks.
-;; - [Kindly](https://github.com/scicloj/kindly) - a tiny library  for specifying the kind of way different things should be viewed
-;; - [Clay](https://github.com/scicloj/clay) - a dynamic workflow for visual exploration & documentation, combining different tools using Kindly
+;; - [Kindly](https://github.com/scicloj/kindly) - a tiny library for specifying the kind of way things should be viewed
+;; - [Clay](https://github.com/scicloj/clay) - a dynamic workflow for visual exploration & documentation
 ;; - [Viz.clj](https://github.com/scicloj/viz.clj) - a (work-in-progress) library for easy data visualizations on top of [Hanami](https://github.com/jsa-aerial/hanami), which is Kindly-aware, and thus fits nicely with Clay
 
 ;; ## Setup
 
-;; For Clay to work, we need to be able to inform it about code evaluations. This requires some setup specific to your editor.
-;; (This is needed only for enjoying Clay's dynamic interaction. For rendering of a whole document, it is not needed.)
-
-;; If your Clojure environment is not supported yet, let us talk and make it work.
+;; For rendering documents like this one with Clay, one only needs to add Clay as a dependency to your project.
+;;
+;; To enjoy Clay's dynamic interaction, it is also needed to inform it about code evaluations. This requires some setup at the editor.
+;;
+;; See some suggested setup for popular editors below. If your favourite editor is not supported yet, let us talk and make it work.
 
 ;; ### Calva
 
@@ -83,157 +84,107 @@
 ;; Now, we can write a namespace and play with Clay.
 
 (ns intro
-  (:require [scicloj.clay.v1.api :as clay]
-            [scicloj.clay.v1.tools :as tools]
-            [scicloj.clay.v1.extensions :as extensions]
-            [scicloj.clay.v1.tool.scittle :as scittle]
-            [scicloj.kindly.v2.api :as kindly]
-            [scicloj.kindly.v2.kind :as kind]
-            [scicloj.kindly.v2.kindness :as kindness]
-            [nextjournal.clerk :as clerk]))
+  (:require [scicloj.clay.v2.api :as clay]
+            [scicloj.kindly.v3.api :as kindly]
+            [scicloj.kindly.v3.kind :as kind]
+            [scicloj.kindly.v3.kindness :as kindness]))
 
-;; Clay can be started with the choice of desired tools to use, as well as extensions that take care of viewing certain datatypes appropriately.
+;; Let us start Clay.
 
-(clay/start! {:tools [tools/clerk
-                      tools/portal
-                      tools/scittle]
-              :extensions [extensions/dataset
-                           extensions/clojisr]})
+(clay/start!)
 
-;; The view of those tools should open automatically (e.g., a Portal window, a Clerk tab in the browser, and a Clay tab with a Scittle-based HTML page).
+;; The browser view should open automatically.
 
 ;; ## A few useful actions
 
-;; Restarting with a new choice of tools and extensions:
+;; Showing the whole namespace:
+(comment
+  (clay/show-doc! "notebooks/intro.clj"))
+
+;; Writing the document:
 
 (comment
-  (clay/restart! {:tools [tools/scittle]
-                  :extensions [extensions/dataset
-                               extensions/clojisr]}))
-
-;; Showing the usual Clerk view of a whole namespace:
-
-(comment
-  (kind/hidden
-   [(clerk/show! "notebooks/intro.clj")]))
-
-;; Static rendering in Clerk:
-
-(comment
-  (clerk/build-static-app! {:paths ["notebooks/intro.clj"]}))
-
-;; Showing the whole namespace using Scittle:
-(comment
-  (scittle/show-doc! "notebooks/intro.clj"))
-
-;; Writing the Scittle-based document:
-
-(comment
-  (do (scittle/show-doc! "notebooks/intro.clj")
-      (scittle/write-html! "docs/index.html")))
-
-;; ## Tools
-
-;; Tools (e.g., `tools/clerk` and `tools/portal` above) are reifications of the `Tool` protocol, that defines common behaviours of tools.
+  (do (clay/show-doc! "notebooks/intro.clj")
+      (clay/write-html! "docs/index.html")))
 
 ;; ## Interaction
 
-;; Clay is not supposed to interfere with the usual way of using tools. One can keep submitting values to Portal, showing files in Clerk, evaluating forms in CIDER or Calva, etc.
-
-;; But Clay adds something: it listens to user evaluations and reflects them visually in the chosen tools.
-
-;; If one has started Clay with `tools/clerk` and `tools/portal`, then after evaluation, the corresponding views (Clerk browser tab, Portal window) should show the evaluation result.
+;; Clay listens to user evaluations and reflects them visually.
 
 (+ 1 2)
 
-;; The way things should be visualized is determined by their so-called kinds.
+;; In Emacs CIDER, after evaluation of a form (or a region),
+;; the browser view should show the evaluation result.
+
+;; In VSCode Calva, a similar effect can be achieved
+;; using the dedicated command and keybinding defined above.
 
 ;; ## Kinds
 
-;; The Kindly library allows one to attach kinds to things. Those kinds can have certain behaviours defined for different tools. Thus, the specified kinds determine the way things are viewed.
+;; The way things should be visualized is determined by the
+;; [Kindly](https://github.com/scicloj/kindly) library.
 
-;; One can check the kind of a value using the `kindly/kind` function.
+;; Clay adopts Kindly's default advice, that will be demonstrated below.
+;; However, user-defined Kindly advices should work as well.
 
-;; There are a few ways to attach a kind to a value.
+;; Kindly advises tools (like Clay) about the kind of way a given context
+;; things should be displayed, by assigning to it a so-called kind.
 
-;; ### Value kind metadata
+;; Please refer to the Kindly documentation for details about specifying
+;; and using kinds.
 
-;; Let us look into this example in [Hiccup](https://github.com/weavejester/hiccup) notation.
+;; ## Examples
 
-(def some-hiccup
-  [:p {:style {:color "#c9a465"}}
-   [:big "hello"]])
+;; ### Plain values
 
-;; We wish to tell Clay that it is of kind `hiccup`.
+;; By default, when there is no kind information provided by Kindly,
+;; values are printed.
 
-;; Let us specify the kind for this value. We'll do it in three ways: by calling a kind, by considering it (using `kindly/consider`), or by considering the corresponding keyword.
+(def people-as-maps
+  (->> (range 29)
+       (mapv (fn [i]
+               {:preferred-language (["clojure" "clojurescript" "babashka"]
+                                     (rand-int 3))
+                :age (rand-int 100)}))))
 
-[(-> some-hiccup kind/hiccup)
- (-> some-hiccup (kindly/consider kind/hiccup))
- (-> some-hiccup (kindly/consider :kind/hiccup))]
+(def people-as-vectors
+  (->> people-as-maps
+       (mapv (juxt :preferred-language :age))))
 
-;; Notice how in all cases, the view is affected accordingly, as we just told Clay to treat the value as Hiccup notation for rendering HTML.
+people-as-maps
 
-;; Let us check the kind in all these cases:
+people-as-vectors
 
-(->> [(-> some-hiccup kind/hiccup)
-      (-> some-hiccup (kindly/consider kind/hiccup))
-      (-> some-hiccup (kindly/consider :kind/hiccup))]
-     (map kindly/kind))
+;; ### Pretty printing
 
-;; Actually, in all these cases, the kind is represeted using metadata attached to the value:
+;; With the the `:kind/pprint` kind, values are pretty-printed.
+(kind/pprint people-as-maps)
 
-(->> [(-> some-hiccup kind/hiccup)
-      (-> some-hiccup (kindly/consider kind/hiccup))
-      (-> some-hiccup (kindly/consider :kind/hiccup))]
-     (map (fn [v]
-            (-> v
-                meta
-                :kindly/kind))))
+(kind/pprint people-as-vectors)
 
-;; ### Code kind metadata
+;; ### Tables
 
-;; Kind metadata can also attached to the code itself (rather than the resulting value).
+;; The `:kind/table` kind can be handy for an interactive table view.
 
-^:kind/hiccup
-some-hiccup
+(kind/table
+ {:column-names [:preferred-language :age]
+  :row-vectors people-as-vectors})
 
-^{:kind/hiccup true}
-some-hiccup
+(kind/table
+ {:column-names [:preferred-language :age]
+  :row-maps people-as-maps})
 
-;; (This option currently does not work in the `clerk` tool.)
 
-;; ### Kindness protocol
-
-;; Another way of specifying kind is implementing the `Kindness` protocol. For example, the Java `BufferedImage` class implements it in order to support viewing images.
-
-(import java.awt.image.BufferedImage)
-
-(->> (BufferedImage. 16 16 BufferedImage/TYPE_INT_RGB)
-     (satisfies? kindness/Kindness))
-
-(-> (BufferedImage. 16 16 BufferedImage/TYPE_INT_RGB)
-    kindly/kind)
-
-;; ## Useful kinds defined in Clay
-
-;; ### Naive
-
-;; The naive kind just uses the usual printing of the value, in any tool.
-
-(-> {:x 9}
-    kind/naive)
 
 ;; ### Hiccup
 
-(-> [:p {:style ; https://www.htmlcsscolor.com/hex/7F5F3F
-         {:color "#7F5F3F"}}
-     "hello"]
-    kind/hiccup)
+;; [Hiccup](https://github.com/weavejester/hiccup), a popular Clojure way to represent HTML, is recognized automatically.
+
+[:p {:style ; https://www.htmlcsscolor.com/hex/7F5F3F
+     {:color "#7F5F3F"}}
+ "hello"]
 
 ;; ### Images
-
-;; Images are handled automatically (technically, this works since the `BufferedImage` class implements the `Kindness` protocol).
 
 (import java.awt.image.BufferedImage
         java.awt.Color
@@ -250,26 +201,13 @@ some-hiccup
 
 (a-piece-of-random-art (+ 40 (rand-int 90)))
 
-;; ### Tables
-
-(-> {:column-names [:preferred-language :age]
-     :row-vectors (for [i (range 99)]
-                    [(["clojure" "clojurescript" "babashka"]
-                      (rand-int 3))
-                     (rand-int 100)])}
-    (kindly/consider kind/table))
-
-(-> {:column-names [:preferred-language :age]
-     :row-maps (for [i (range 99)]
-                 {:preferred-language (["clojure" "clojurescript" "babashka"]
-                                       (rand-int 3))
-                  :age (rand-int 100)})}
-    (kindly/consider kind/table))
-
 ;; ### Datasets
-;; For [tech.ml.dataset](https://github.com/techascent/tech.ml.dataset) datasets to render coodectly, it is necessary to use the `dataset` extension when starting Clay (see above), and to have tech.ml.dataset as a dependency of the project.
 
-;; In this example, let us create a dataset using [Tablecloth](https://github.com/scicloj/tablecloth).
+;; [tech.ml.dataset](https://github.com/techascent/tech.ml.dataset) datasets use the default
+;; printing of the library, which is then rendered as Markdown
+;; (with some CSS styling).
+
+;; Let us create such a dataset using [Tablecloth](https://github.com/scicloj/tablecloth).
 
 (require '[tablecloth.api :as tc])
 
@@ -279,26 +217,18 @@ some-hiccup
 
 ;; #### Known issues
 
-;; With the current Markdown implementation used by `tool/scittle` (based on [Cybermonday](https://github.com/kiranshila/cybermonday)), brackets inside datasets cells are not visible.
+;; With the current Markdown implementation, used by Clay (based on [Cybermonday](https://github.com/kiranshila/cybermonday)), brackets inside datasets cells are not visible.
 
 (-> {:x [1 [2 3] 4]
      :y [:A :B :C]}
     tc/dataset)
 
-;; For now, cases of this kind can be handled by the user by naive rendering of the printed output:
+;; For now, cases of this kind can be handled by the user by switching to the `:kind/pprint` kind.
 
 (-> {:x [1 [2 3] 4]
      :y [:A :B :C]}
     tc/dataset
-    kind/naive)
-
-;; ### RObjects (ClijisR)
-
-;; For [ClojisR](https://github.com/scicloj/clojisr) `RObjects` (which are Clojure handles to objects of the R language) to render coodectly, it is necessary to use the `clojisr` extension when starting Clay (see above), and to have ClojisR as a dependency of the project.
-
-(require '[clojisr.v1.r :as r])
-
-(r/r '(rnorm 9))
+    kind/pprint)
 
 ;; ### [Vega](https://vega.github.io/vega/) and [Vega-Lite](https://vega.github.io/vega-lite/)
 
@@ -310,7 +240,7 @@ some-hiccup
         :x {:field "x", :type "quantitative"},
         :y {:field "y", :type "quantitative"},
         :fill {:field "z", :type "nominal"}}}
-      (kindly/consider kind/vega)))
+      kind/vega))
 
 (defn random-data [n]
   (->> (repeatedly n #(- (rand) 0.5))
@@ -384,8 +314,7 @@ some-hiccup
   (+ 1 2))
 
 (delay
-  (-> [:div [:big "hi......."]]
-      (kindly/consider kind/hiccup)))
+  [:div [:big "hi......."]])
 
 ;; ## Tests
 
@@ -402,21 +331,6 @@ some-hiccup
 ;; We are considering a so-called "doctest" setup involving such checks, so that actual Clojure tests would be derived automatically from them.
 
 ;; This would open the way for literate testing / testable documentation solutions, such as those we have been using in the past (e.g., in [tutorials](https://scicloj.github.io/clojisr/doc/clojisr/v1/tutorial-test/) of ClojisR using Notespace v2).
-
-;; ## Nesting
-
-;; Different kinds of views should (eventually) nest correctly, thanks to the nesting support of Portal, Clerk, etc.
-
-;; Here are a few Vega-Lite specs (using the function defined above) inside a Hiccup block:
-
-(->> [10 100 1000]
-     (map (fn [n]
-            [:div {:style {:width "400px"}}
-             [:big (str "n=" n)]
-             (random-vega-plot n)]))
-     (into [:div])
-     kind/hiccup)
-
 
 ;; ## Development
 
@@ -441,15 +355,5 @@ some-hiccup
 ;; ```
 
 ;; Your library will be deployed to org.scicloj/clay on clojars.org by default.
-
-;; ## Planned features
-
-;; - a quick-start project template
-
-;; - client-server interactive widgets (for a proof-of-concept, see [the 2022-04-16 example project](https://github.com/scicloj/clay/tree/main/examples/example-project-20220416) and the [Clojure-Asia talk](https://www.youtube.com/watch?v=gFNPtgAw36k).)
-
-;; - presentation slides
-
-;; - more kinds of visualizations
 
 :bye