+++ src/scicloj/clay/v2/tool.clj
@@ -1,0 +1,7 @@
+(ns scicloj.clay.v2.tool)
+
+(defprotocol Tool
+  (setup! [this config])
+  (open! [this])
+  (close! [this])
+  (show! [this context]))