+++ src/scicloj/clay/v2/tools.clj
@@ -1,0 +1,5 @@
+(ns scicloj.clay.v2.tools
+  (:require
+   [scicloj.clay.v2.tool.scittle :as tool.scittle]))
+
+(def scittle tool.scittle/tool)