+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 All notable changes to this project will be documented in this file. This change log follows the conventions of [keepachangelog.com](http://keepachangelog.com/).
 
+## [2-alpha1]
+- rendeding documents without relying on Clerk
+- switched to clay.v2 namespaces
+- switched to kindly v3
+
 ## [1-alpha15] - 2022-08-07
 - catching errors
 - pretty printing where appropriate